+++ src/components/dashboard/templating/message/types.ts
@@ -1,0 +1,16 @@
+export interface Message {
+    embeds: Embed[];
+    content: string;
+}
+
+export interface Embed {
+    title: string;
+    description: string;
+    fields: EmbedField[];
+}
+
+export interface EmbedField {
+    name: string;
+    value: string;
+    inline: boolean;
+}
