+++ src/components/inputs/generic/InputDispatcher.svelte
@@ -18,8 +18,8 @@ Note: this may be less performant than using the concrete input components direc
 	import InputDescription from '../InputDescription.svelte';
 	import RoleInput from '../RoleInput.svelte';
 	import Modifier from '../Modifier.svelte';
-	import TemplateChooser from '../../dashboard/templatechooser/TemplateChooser.svelte';
-	import TemplateBuilder from '../../dashboard/message_templatebuilder/TemplateBuilder.svelte';
+	import TemplateChooser from '../../dashboard/templating/TemplateChooser.svelte';
+	import TemplateBuilder from '../../dashboard/templating/TemplateBuilder.svelte';
 
 	export let type: string;
 
@@ -167,7 +167,7 @@ Note: this may be less performant than using the concrete input components direc
 	/>
 {:else if type == 'string:textarea:template'}
 	<Label {id} {label} />
-	<TemplateBuilder bind:rawTemplateOutput={value} />
+	<TemplateBuilder bind:output={value} />
 	<small class="text-gray-500 dark:text-gray-400"
 		>See our documentation to learn more about templating</small
 	>
